+++ engine-standalone-tracing/Cargo.toml
@@ -2,7 +2,7 @@
 name = "engine-standalone-tracing"
 version = "0.1.0"
 edition = "2018"
-authors = ["Aurora <hello@aurora.dev>"]
+authors = ["Aurora Labs <hello@aurora.dev>"]
 description = "Aurora engine standalone tracing library. Provides functions and types for extracing geth-like traces from standalone engine execution."
 homepage = "https://github.com/aurora-is-near/aurora-engine"
 repository = "https://github.com/aurora-is-near/aurora-engine"
