+++ engine-tests/src/benches/nft_pagination.rs
@@ -1,6 +1,6 @@
 use crate::prelude::{Address, Wei, U256};
 use crate::test_utils::{self, solidity};
-use aurora_engine::transaction::legacy::TransactionLegacy;
+use aurora_engine_transactions::legacy::TransactionLegacy;
 use secp256k1::SecretKey;
 use std::path::{Path, PathBuf};
 use std::process::Command;
@@ -54,6 +54,7 @@ pub(crate) fn measure_gas_usage(
 }
 
 struct MarketPlaceConstructor(solidity::ContractConstructor);
+
 struct MarketPlace(solidity::DeployedContract);
 
 impl MarketPlaceConstructor {
