+++ engine-tests/src/prelude.rs
@@ -4,12 +4,13 @@ mod v0 {
     #[cfg(feature = "meta-call")]
     pub use aurora_engine::meta_parsing;
     pub use aurora_engine::parameters;
-    pub use aurora_engine::transaction;
     pub use aurora_engine_sdk as sdk;
+    pub use aurora_engine_transactions as transactions;
     pub use aurora_engine_types::parameters::*;
     pub use aurora_engine_types::storage;
     pub use aurora_engine_types::types::*;
     pub use aurora_engine_types::*;
     pub use borsh::{BorshDeserialize, BorshSerialize};
 }
+
 pub use v0::*;
