+++ engine-tests/src/test_utils/solidity.rs
@@ -1,4 +1,4 @@
-use crate::prelude::{transaction::legacy::TransactionLegacy, Address, U256};
+use crate::prelude::{transactions::legacy::TransactionLegacy, Address, U256};
 use near_sdk::serde_json;
 use serde::Deserialize;
 use std::fs;
