+++ engine-tests/src/test_utils/standalone/mod.rs
@@ -1,7 +1,7 @@
 use aurora_engine::engine;
 use aurora_engine::parameters::{CallArgs, DeployErc20TokenArgs, SubmitResult, TransactionStatus};
-use aurora_engine::transaction::legacy::{LegacyEthSignedTransaction, TransactionLegacy};
 use aurora_engine_sdk::env::{self, Env};
+use aurora_engine_transactions::legacy::{LegacyEthSignedTransaction, TransactionLegacy};
 use aurora_engine_types::types::{Address, NearGas, Wei};
 use aurora_engine_types::{H256, U256};
 use borsh::BorshDeserialize;
