+++ engine-tests/src/tests/one_inch.rs
@@ -125,7 +125,7 @@ fn deploy_1_inch_limit_order_contract(
         test_utils::solidity::ContractConstructor::compile_from_extended_json(contract_path);
 
     let nonce = signer.use_nonce();
-    let deploy_tx = crate::prelude::transaction::legacy::TransactionLegacy {
+    let deploy_tx = crate::prelude::transactions::legacy::TransactionLegacy {
         nonce: nonce.into(),
         gas_price: Default::default(),
         gas_limit: u64::MAX.into(),
