+++ engine-transactions/src/eip_1559.rs
@@ -1,34 +1,22 @@
-use crate::prelude::precompiles::secp256k1::ecrecover;
-use crate::prelude::{sdk, Address, Vec, Wei, H160, H256, U256};
+use crate::eip_2930::AccessTuple;
+use aurora_engine_precompiles::secp256k1::ecrecover;
+use aurora_engine_types::types::{Address, Wei};
+use aurora_engine_types::{Vec, U256};
 use rlp::{Decodable, DecoderError, Encodable, Rlp, RlpStream};
 
-/// Type indicator (per EIP-2718) for access list transactions
-pub const TYPE_BYTE: u8 = 0x01;
+/// Type indicator (per EIP-1559)
+pub const TYPE_BYTE: u8 = 0x02;
 
+/// A EIP-1559 transaction kind from the London hard fork.
+///
+/// See [EIP-1559](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1559.md)
+/// for more details.
 #[derive(Debug, Eq, PartialEq, Clone)]
-pub struct AccessTuple {
-    pub address: H160,
-    pub storage_keys: Vec<H256>,
-}
-
-impl Decodable for AccessTuple {
-    fn decode(rlp: &Rlp<'_>) -> Result<Self, DecoderError> {
-        let address = rlp.val_at(0)?;
-        let storage_keys = rlp.list_at(1)?;
-
-        Ok(Self {
-            address,
-            storage_keys,
-        })
-    }
-}
-
-/// See https://eips.ethereum.org/EIPS/eip-2930
-#[derive(Debug, Eq, PartialEq, Clone)]
-pub struct Transaction2930 {
+pub struct Transaction1559 {
     pub chain_id: u64,
     pub nonce: U256,
-    pub gas_price: U256,
+    pub max_priority_fee_per_gas: U256,
+    pub max_fee_per_gas: U256,
     pub gas_limit: U256,
     pub to: Option<Address>,
     pub value: Wei,
@@ -36,22 +24,23 @@ pub struct Transaction2930 {
     pub access_list: Vec<AccessTuple>,
 }
 
-impl Transaction2930 {
+impl Transaction1559 {
     /// RLP encoding of the data for an unsigned message (used to make signature)
     pub fn rlp_append_unsigned(&self, s: &mut RlpStream) {
-        self.rlp_append(s, 8);
+        self.rlp_append(s, 9);
     }
 
     /// RLP encoding for a signed message (used to encode the transaction for sending to tx pool)
     pub fn rlp_append_signed(&self, s: &mut RlpStream) {
-        self.rlp_append(s, 11);
+        self.rlp_append(s, 12);
     }
 
     fn rlp_append(&self, s: &mut RlpStream, list_len: usize) {
         s.begin_list(list_len);
         s.append(&self.chain_id);
         s.append(&self.nonce);
-        s.append(&self.gas_price);
+        s.append(&self.max_priority_fee_per_gas);
+        s.append(&self.max_fee_per_gas);
         s.append(&self.gas_limit);
         match self.to.as_ref() {
             None => s.append(&""),
@@ -72,20 +61,20 @@ impl Transaction2930 {
 }
 
 #[derive(Debug, Eq, PartialEq, Clone)]
-pub struct SignedTransaction2930 {
-    pub transaction: Transaction2930,
+pub struct SignedTransaction1559 {
+    pub transaction: Transaction1559,
     /// The parity (0 for even, 1 for odd) of the y-value of a secp256k1 signature.
     pub parity: u8,
     pub r: U256,
     pub s: U256,
 }
 
-impl SignedTransaction2930 {
+impl SignedTransaction1559 {
     pub fn sender(&self) -> Option<Address> {
         let mut rlp_stream = RlpStream::new();
         rlp_stream.append(&TYPE_BYTE);
         self.transaction.rlp_append_unsigned(&mut rlp_stream);
-        let message_hash = sdk::keccak(rlp_stream.as_raw());
+        let message_hash = aurora_engine_sdk::keccak(rlp_stream.as_raw());
         ecrecover(
             message_hash,
             &super::vrs_to_arr(self.parity, self.r, self.s),
@@ -94,7 +83,7 @@ impl SignedTransaction2930 {
     }
 }
 
-impl Encodable for SignedTransaction2930 {
+impl Encodable for SignedTransaction1559 {
     fn rlp_append(&self, s: &mut RlpStream) {
         self.transaction.rlp_append_signed(s);
         s.append(&self.parity);
@@ -103,27 +92,29 @@ impl Encodable for SignedTransaction2930 {
     }
 }
 
-impl Decodable for SignedTransaction2930 {
+impl Decodable for SignedTransaction1559 {
     fn decode(rlp: &Rlp<'_>) -> Result<Self, DecoderError> {
-        if rlp.item_count() != Ok(11) {
+        if rlp.item_count() != Ok(12) {
             return Err(rlp::DecoderError::RlpIncorrectListLen);
         }
         let chain_id = rlp.val_at(0)?;
         let nonce = rlp.val_at(1)?;
-        let gas_price = rlp.val_at(2)?;
-        let gas_limit = rlp.val_at(3)?;
-        let to = super::rlp_extract_to(rlp, 4)?;
-        let value = Wei::new(rlp.val_at(5)?);
-        let data = rlp.val_at(6)?;
-        let access_list = rlp.list_at(7)?;
-        let parity = rlp.val_at(8)?;
-        let r = rlp.val_at(9)?;
-        let s = rlp.val_at(10)?;
+        let max_priority_fee_per_gas = rlp.val_at(2)?;
+        let max_fee_per_gas = rlp.val_at(3)?;
+        let gas_limit = rlp.val_at(4)?;
+        let to = super::rlp_extract_to(rlp, 5)?;
+        let value = Wei::new(rlp.val_at(6)?);
+        let data = rlp.val_at(7)?;
+        let access_list = rlp.list_at(8)?;
+        let parity = rlp.val_at(9)?;
+        let r = rlp.val_at(10)?;
+        let s = rlp.val_at(11)?;
         Ok(Self {
-            transaction: Transaction2930 {
+            transaction: Transaction1559 {
                 chain_id,
                 nonce,
-                gas_price,
+                max_priority_fee_per_gas,
+                max_fee_per_gas,
                 gas_limit,
                 to,
                 value,
