+++ engine/src/prelude.rs
@@ -2,6 +2,7 @@ mod v0 {
     pub use aurora_engine_precompiles as precompiles;
     pub use aurora_engine_sdk as sdk;
     pub use aurora_engine_sdk::types::*;
+    pub use aurora_engine_transactions as transactions;
     pub use aurora_engine_types::account_id::*;
     pub use aurora_engine_types::parameters::*;
     pub use aurora_engine_types::storage::*;
@@ -9,4 +10,5 @@ mod v0 {
     pub use aurora_engine_types::*;
     pub use borsh::{BorshDeserialize, BorshSerialize};
 }
+
 pub use v0::*;
